Moved two handlers to separate files
