+++ lambda/custom/handlers/session-ended-handler.js
@@ -1,0 +1,12 @@
+const Alexa = require('ask-sdk-core');
+
+export const SessionEndedRequestHandler = {
+    canHandle(handlerInput) {
+      return handlerInput.requestEnvelope.request.type === 'SessionEndedRequest';
+    },
+    handle(handlerInput) {
+      console.log(`Session ended with reason: ${handlerInput.requestEnvelope.request.reason}`);
+  
+      return handlerInput.responseBuilder.getResponse();
+    },
+  };
